Update README.md
The lab involves setting up a WordPress environment on an Ubuntu 20.04 virtual machine. The process requires installing and configuring necessary parts like Apache, PHP, and MySQL, as well as safeguarding the set up using SSL certificates. The goal is to produce a fully functional WordPress site that can be both secure and accessible.